Used real data on projects carousel

+++ feature/home/src/main/java/com/devalr/home/interactions/Action.kt
@@ -3,5 +3,5 @@ package com.devalr.home.interactions
 sealed interface Action {
     data object OnAppear : Action
     data object OnStartPainting : Action
-    data class OnOpenProjectDetail(val projectId: Int) : Action
+    data class OnOpenProjectDetail(val projectId: Long) : Action
 }
