+++ feature/home/src/main/java/com/devalr/home/interactions/Event.kt
@@ -2,5 +2,5 @@ package com.devalr.home.interactions
 
 sealed interface Event {
     data object LaunchStartPaintModal : Event
-    data class NavigateToProject(val projectId: Int) : Event
+    data class NavigateToProject(val projectId: Long) : Event
 }
